move EventCards to separate component, allow filtering for globalEvents view
