+++ packages/lesswrong/components/events/EventsHome.tsx
@@ -1,26 +1,21 @@
 import { Components, registerComponent, } from '../../lib/vulcan-lib';
 import React, { useState, useEffect } from 'react';
-import { Link } from '../../lib/reactRouterWrapper';
 import { userGetLocation } from '../../lib/collections/users/helpers';
 import { useCurrentUser } from '../common/withUser';
 import { createStyles } from '@material-ui/core/styles';
-import NotificationsNoneIcon from '@material-ui/icons/NotificationsNone';
-import { useLocation } from '../../lib/routeUtil';
-import { useDialog } from '../common/withDialog'
 import * as _ from 'underscore';
+import NotificationsNoneIcon from '@material-ui/icons/NotificationsNone';
+import { useDialog } from '../common/withDialog'
 import {AnalyticsContext} from "../../lib/analyticsEvents";
 import { forumTypeSetting } from '../../lib/instanceSettings';
 import { useUpdate } from '../../lib/crud/withUpdate';
 import { pickBestReverseGeocodingResult } from '../../server/mapsUtils';
 import { useGoogleMaps } from '../form-components/LocationFormComponent';
-import { Button, Card, CardContent, CardMedia, CircularProgress, Select, MenuItem } from '@material-ui/core';
+import { Button, CircularProgress, Select, MenuItem } from '@material-ui/core';
 import { useMulti } from '../../lib/crud/withMulti';
-import { prettyEventDateTimes } from '../../lib/collections/posts/helpers';
-import { useTimezone } from '../common/withTimezone';
 import { getBrowserLocalStorage } from '../async/localStorageHandlers';
 import { geoSuggestStyles } from '../form-components/LocationFormComponent'
 import Geosuggest from 'react-geosuggest';
-import { cloudinaryCloudNameSetting } from '../../lib/publicSettings';
 
 const styles = createStyles((theme: ThemeType): JssStyles => ({
   section: {
@@ -97,94 +92,6 @@ const styles = createStyles((theme: ThemeType): JssStyles => ({
       gridTemplateColumns: 'auto',
     }
   },
-  eventCard: {
-    position: 'relative',
-    width: 373,
-    height: 374,
-    borderRadius: 0,
-    overflow: 'visible',
-    [theme.breakpoints.down('xs')]: {
-      // width: 'auto',
-      height: 370
-    }
-  },
-  eventCardContent: {
-    position: 'relative',
-    height: 170,
-    // display: 'grid',
-    // gridTemplateAreas: `
-    //   "time ."
-    //   "title ."
-    //   "location ."
-    //   "group tag"
-    // `,
-    // gridGap: '8px',
-    // gridTemplateRows: '18px 60px 18px 18px',
-    // alignItems: 'baseline',
-    // [theme.breakpoints.down('xs')]: {
-    //   gridTemplateAreas: `
-    //   "time"
-    //   "title"
-    //   "location"
-    //   "group"
-    // `,
-    // }
-  },
-  eventCardTime: {
-    ...theme.typography.commentStyle,
-    gridArea: 'time',
-    fontSize: 14,
-    color: theme.palette.primary.main
-  },
-  eventCardTitle: {
-    ...theme.typography.headline,
-    gridArea: 'title',
-    fontSize: 20,
-    display: '-webkit-box',
-    "-webkit-line-clamp": 2,
-    "-webkit-box-orient": 'vertical',
-    overflow: 'hidden',
-    marginTop: 8,
-    marginBottom: 0
-  },
-  eventCardLocation: {
-    ...theme.typography.commentStyle,
-    gridArea: 'location',
-    // textAlign: 'right',
-    color: "rgba(0, 0, 0, 0.7)",
-    fontSize: 14,
-    textOverflow: 'ellipsis',
-    whiteSpace: 'nowrap',
-    overflow: 'hidden',
-    marginTop: 10,
-    // [theme.breakpoints.down('xs')]: {
-    //   textAlign: 'left'
-    // }
-  },
-  eventCardGroup: {
-    ...theme.typography.commentStyle,
-    gridArea: 'group',
-    maxWidth: 290,
-    fontStyle: 'italic',
-    color: "rgba(0, 0, 0, 0.5)",
-    fontSize: 14,
-    textOverflow: 'ellipsis',
-    whiteSpace: 'nowrap',
-    overflow: 'hidden',
-    marginTop: 10,
-  },
-  eventCardTag: {
-    ...theme.typography.commentStyle,
-    position: 'absolute',
-    bottom: 22,
-    right: 22,
-    gridArea: 'tag',
-    // textAlign: 'right',
-    fontSize: 14,
-    [theme.breakpoints.down('xs')]: {
-      display: 'none'
-    }
-  },
   loadMoreRow: {
     gridColumnStart: 1,
     gridColumnEnd: -1,
@@ -215,52 +122,12 @@ const styles = createStyles((theme: ThemeType): JssStyles => ({
   },
 }))
 
-/**
- * Randomly assigns one of twelve images to the event based on its _id.
- * 
- * @param eventId - _id of the event/post
- * @returns img url
- */
-const randomEventImg = (eventId) => {
-  const num = _.range(eventId.length).reduce((prev, next) => {
-    return prev + eventId.charCodeAt(next)
-  }, 0) % 12
-
-  switch (num) {
-    case 0:
-      return 'Banner/c4xfkjbyrkgzk67j8yhx'
-    case 1:
-      return 'Banner/lm7in6trshkcnqgeybqr'
-    case 2:
-      return 'Banner/tpa8dburf2fpv7vkqw3h'
-    case 3:
-      return 'Banner/k1uurxviebati6mpaund'
-    case 4:
-      return 'Banner/mngmri7qblzit9jigof4'
-    case 5:
-      return 'Banner/rpeoujevvhdhulgevv3u'
-    case 6:
-      return 'Banner/vxquzxthtaiha6r5lzbq'
-    case 7:
-      return 'Banner/rp0ywuja8mflliboqoxb'
-    case 8:
-      return 'Banner/yvfw6msbycjpz7wncawu'
-    case 9:
-      return 'Banner/rpsvlou1aci2rpz0zpfs'
-    case 10:
-      return 'Banner/jouhj45hrkkfkhinknbg'
-    default:
-      return 'Banner/qfffq3yggslcyttsonxq'
-  }
-}
 
 const EventsHome = ({classes}: {
   classes: ClassesType,
 }) => {
   const currentUser = useCurrentUser();
   const { openDialog } = useDialog();
-  const { query } = useLocation();
-  const { timezone } = useTimezone()
   
   const [userLocation, setUserLocation] = useState(() => {
     if (currentUser) {
@@ -363,12 +230,7 @@ const EventsHome = ({classes}: {
     });
   }
   
-  const getEventLocation = (event) => {
-    if (event.onlineEvent) return 'Online'
-    return event.location ? event.location.slice(0, event.location.lastIndexOf(',')) : ''
-  }
-  
-  const { SingleColumnSection, SectionTitle, SectionFooter, Typography, SectionButton, AddToCalendarIcon, EventTime, Loading, PostsItemTooltipWrapper, CloudinaryImage2, HighlightedEventCard } = Components
+  const { HighlightedEventCard, EventCards } = Components
 
   const filters: PostsViewTerms = {}
   if (placeFilter === 'in-person') {
@@ -387,7 +249,7 @@ const EventsHome = ({classes}: {
     ...filters,
   }
   
-  const { results, loading, error, showLoadMore, loadMore } = useMulti({
+  const { results, loading, showLoadMore, loadMore } = useMulti({
     terms: eventsListTerms,
     collectionName: "Posts",
     fragmentName: 'PostsList',
@@ -411,7 +273,7 @@ const EventsHome = ({classes}: {
   let loadMoreButton = showLoadMore && <button className={classes.loadMore} onClick={() => loadMore(null)}>
     Load More
   </button>
-  if (loading && results) {
+  if (loading && results?.length) {
     loadMoreButton = <CircularProgress size={16} />
   }
 
@@ -460,34 +322,7 @@ const EventsHome = ({classes}: {
               </Select>
             </div>
 
-            {results ? results.map(event => {
-              return <Card key={event._id} className={classes.eventCard}>
-                {/* <CloudinaryImage2 height={200} width={373} publicId={event.eventImageId || randomEventImg(event._id)} /> */}
-                <CloudinaryImage2 height={200} width={373} publicId={event.eventImageId || 'Banner/yeldubyolqpl3vqqy0m6'} />
-                <CardContent className={classes.eventCardContent}>
-                  <div className={classes.eventCardTime}>
-                    {prettyEventDateTimes(event, timezone, true)}
-                  </div>
-                  <PostsItemTooltipWrapper
-                    post={event}
-                    className={''}
-                  >
-                    <div className={classes.eventCardTitle}>
-                      <Link to={`/events/${event._id}/${event.slug}`}>{event.title}</Link>
-                    </div>
-                  </PostsItemTooltipWrapper>
-                  <div className={classes.eventCardLocation}>{getEventLocation(event)}</div>
-                  {event.group && <div className={classes.eventCardGroup} title={event.group.name}>
-                    <Link to={`/groups/${event.group._id}`}>{event.group.name}</Link>
-                  </div>}
-                  <div className={classes.eventCardTag}>
-                    <AddToCalendarIcon post={event} />
-                  </div>
-                </CardContent>
-              </Card>
-            }) : _.range(6).map((i) => {
-              return <Card key={i} className={classes.eventCard}></Card>
-            })}
+            <EventCards events={results} loading={loading} numDefaultCards={6} />
             
             <div className={classes.loadMoreRow}>
               <Button variant="text" color="primary" onClick={openEventNotificationsForm} className={classes.eventNotificationsBtn}>
